# swagger.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/api/v1":{"get":{"summary":"Home","operationId":"home_api_v1_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/users":{"get":{"summary":"Get Users","operationId":"get_users_api_v1_users_get","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string","title":"Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/accounts":{"get":{"summary":"Get Accounts","operationId":"get_accounts_api_v1_accounts_get","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string","title":"Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/transactions":{"get":{"summary":"Get Transactions","operationId":"get_transactions_api_v1_transactions_get","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string","title":"Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/transfer":{"post":{"summary":"Transfer Money","operationId":"transfer_money_api_v1_transfer_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_transfer_money_api_v1_transfer_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/withdraw":{"post":{"summary":"Withdraw Money","operationId":"withdraw_money_api_v1_withdraw_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_withdraw_money_api_v1_withdraw_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/login":{"post":{"summary":"Login User","operationId":"login_user_api_v1_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_login_user_api_v1_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_login_user_api_v1_login_post":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"Body_login_user_api_v1_login_post"},"Body_transfer_money_api_v1_transfer_post":{"properties":{"token":{"type":"string","title":"Token"},"from_account_id":{"type":"integer","title":"From Account Id"},"to_account_id":{"type":"integer","title":"To Account Id"},"amount":{"type":"number","title":"Amount"}},"type":"object","required":["token","from_account_id","to_account_id","amount"],"title":"Body_transfer_money_api_v1_transfer_post"},"Body_withdraw_money_api_v1_withdraw_post":{"properties":{"token":{"type":"string","title":"Token"},"account_id":{"type":"integer","title":"Account Id"},"amount":{"type":"number","title":"Amount"}},"type":"object","required":["token","account_id","amount"],"title":"Body_withdraw_money_api_v1_withdraw_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}